refactor to minimize start bundle
