+++ x-pack/plugins/aiops/public/components/explain_log_rate_spikes/explain_log_rate_spikes_wrapper.tsx
@@ -5,36 +5,29 @@
  * 2.0.
  */
 
-import React, { FC, useCallback, useEffect, useState } from 'react';
+import React, { FC, useCallback, useEffect } from 'react';
 import { Filter, Query } from '@kbn/es-query';
 import { i18n } from '@kbn/i18n';
+import { parse, stringify } from 'query-string';
+import { isEqual } from 'lodash';
+import { encode } from 'rison-node';
+import { useHistory, useLocation } from 'react-router-dom';
 
-import {
-  EuiFlexGroup,
-  EuiFlexItem,
-  EuiHorizontalRule,
-  EuiPageBody,
-  EuiPageContentBody,
-  EuiPageContentHeader,
-  EuiPageContentHeaderSection,
-  EuiSpacer,
-  EuiTitle,
-} from '@elastic/eui';
-
-import type { WindowParameters } from '@kbn/aiops-utils';
-import type { DataView } from '@kbn/data-views-plugin/public';
-
-import { useData } from '../../hooks/use_data';
-import { useUrlState, usePageUrlState, AppStateKey } from '../../hooks/url_state';
-
-import { FullTimeRangeSelector } from '../full_time_range_selector';
-import { DocumentCountContent } from '../document_count_content/document_count_content';
-import { DatePickerWrapper } from '../date_picker_wrapper';
+import { EuiPageBody } from '@elastic/eui';
+import { DataView } from '@kbn/data-views-plugin/public';
 
 import { ExplainLogRateSpikes } from './explain_log_rate_spikes';
-import { SearchPanel } from '../search_panel';
 import { SEARCH_QUERY_LANGUAGE, SearchQueryLanguage } from '../../../common/types';
 import { useAiOpsKibana } from '../../kibana_context';
+import {
+  Accessor,
+  Dictionary,
+  parseUrlState,
+  Provider as UrlStateContextProvider,
+  isRisonSerializationRequired,
+  getNestedProperty,
+  SetUrlState,
+} from '../../hooks/url_state';
 
 export interface ExplainLogRateSpikesWrapperProps {
   /** The data view to analyze. */
@@ -62,15 +55,15 @@ export const getDefaultAiOpsListState = (
   ...overrides,
 });
 
-const restorableDefaults = getDefaultAiOpsListState();
+export const restorableDefaults = getDefaultAiOpsListState();
 
 export const ExplainLogRateSpikesWrapper: FC<ExplainLogRateSpikesWrapperProps> = ({ dataView }) => {
   const { services } = useAiOpsKibana();
-  const { notifications, data } = services;
+  const { notifications } = services;
   const { toasts } = notifications;
 
-  const [aiopsListState, setAiopsListState] = usePageUrlState(AppStateKey, restorableDefaults);
-  const [globalState, setGlobalState] = useUrlState('_g');
+  const history = useHistory();
+  const { search: urlSearchString } = useLocation();
 
   useEffect(() => {
     if (!dataView.isTimeBased()) {
@@ -89,135 +82,75 @@ export const ExplainLogRateSpikesWrapper: FC<ExplainLogRateSpikesWrapperProps> =
     }
   }, [dataView, toasts]);
 
-  const setSearchParams = useCallback(
-    (searchParams: {
-      searchQuery: Query['query'];
-      searchString: Query['query'];
-      queryLanguage: SearchQueryLanguage;
-      filters: Filter[];
-    }) => {
-      setAiopsListState({
-        ...aiopsListState,
-        searchQuery: searchParams.searchQuery,
-        searchString: searchParams.searchString,
-        searchQueryLanguage: searchParams.queryLanguage,
-        filters: searchParams.filters,
-      });
+  if (!dataView) return null;
+
+  const setUrlState: SetUrlState = useCallback(
+    (
+      accessor: Accessor,
+      attribute: string | Dictionary<any>,
+      value?: any,
+      replaceState?: boolean
+    ) => {
+      const prevSearchString = urlSearchString;
+      const urlState = parseUrlState(prevSearchString);
+      const parsedQueryString = parse(prevSearchString, { sort: false });
+
+      if (!Object.prototype.hasOwnProperty.call(urlState, accessor)) {
+        urlState[accessor] = {};
+      }
+
+      if (typeof attribute === 'string') {
+        if (isEqual(getNestedProperty(urlState, `${accessor}.${attribute}`), value)) {
+          return prevSearchString;
+        }
+
+        urlState[accessor][attribute] = value;
+      } else {
+        const attributes = attribute;
+        Object.keys(attributes).forEach((a) => {
+          urlState[accessor][a] = attributes[a];
+        });
+      }
+
+      try {
+        const oldLocationSearchString = stringify(parsedQueryString, {
+          sort: false,
+          encode: false,
+        });
+
+        Object.keys(urlState).forEach((a) => {
+          if (isRisonSerializationRequired(a)) {
+            parsedQueryString[a] = encode(urlState[a]);
+          } else {
+            parsedQueryString[a] = urlState[a];
+          }
+        });
+        const newLocationSearchString = stringify(parsedQueryString, {
+          sort: false,
+          encode: false,
+        });
+
+        if (oldLocationSearchString !== newLocationSearchString) {
+          const newSearchString = stringify(parsedQueryString, { sort: false });
+          if (replaceState) {
+            history.replace({ search: newSearchString });
+          } else {
+            history.push({ search: newSearchString });
+          }
+        }
+      } catch (error) {
+        // eslint-disable-next-line no-console
+        console.error('Could not save url state', error);
+      }
     },
-    [aiopsListState, setAiopsListState]
+    [history, urlSearchString]
   );
 
-  const { docStats, timefilter, earliest, latest, searchQueryLanguage, searchString, searchQuery } =
-    useData(dataView, aiopsListState, setGlobalState);
-
-  useEffect(() => {
-    return () => {
-      // When navigating away from the index pattern
-      // Reset all previously set filters
-      // to make sure new page doesn't have unrelated filters
-      data.query.filterManager.removeAll();
-    };
-  }, [dataView.id, data.query.filterManager]);
-
-  const [windowParameters, setWindowParameters] = useState<WindowParameters | undefined>();
-
-  useEffect(() => {
-    if (globalState?.time !== undefined) {
-      timefilter.setTime({
-        from: globalState.time.from,
-        to: globalState.time.to,
-      });
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify(globalState?.time), timefilter]);
-
-  useEffect(() => {
-    if (globalState?.refreshInterval !== undefined) {
-      timefilter.setRefreshInterval(globalState.refreshInterval);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [JSON.stringify(globalState?.refreshInterval), timefilter]);
-
-  useEffect(() => {
-    // Update data query manager if input string is updated
-    data?.query.queryString.setQuery({
-      query: searchString,
-      language: searchQueryLanguage,
-    });
-  }, [data, searchQueryLanguage, searchString]);
-
-  if (!dataView || !timefilter) return null;
-
   return (
-    <EuiPageBody data-test-subj="aiopsIndexPage" paddingSize="none" panelled={false}>
-      <EuiFlexGroup gutterSize="m">
-        <EuiFlexItem>
-          <EuiPageContentHeader className="aiopsPageHeader">
-            <EuiPageContentHeaderSection>
-              <div className="aiopsTitleHeader">
-                <EuiTitle size={'s'}>
-                  <h2>{dataView.title}</h2>
-                </EuiTitle>
-              </div>
-            </EuiPageContentHeaderSection>
-
-            <EuiFlexGroup
-              alignItems="center"
-              justifyContent="flexEnd"
-              gutterSize="s"
-              data-test-subj="aiopsTimeRangeSelectorSection"
-            >
-              {dataView.timeFieldName !== undefined && (
-                <EuiFlexItem grow={false}>
-                  <FullTimeRangeSelector
-                    dataView={dataView}
-                    query={undefined}
-                    disabled={false}
-                    timefilter={timefilter}
-                  />
-                </EuiFlexItem>
-              )}
-              <EuiFlexItem grow={false}>
-                <DatePickerWrapper />
-              </EuiFlexItem>
-            </EuiFlexGroup>
-          </EuiPageContentHeader>
-        </EuiFlexItem>
-      </EuiFlexGroup>
-      <EuiHorizontalRule />
-      <EuiPageContentBody>
-        <EuiFlexGroup gutterSize="m" direction="column">
-          <EuiFlexItem>
-            <SearchPanel
-              dataView={dataView}
-              searchString={searchString}
-              searchQuery={searchQuery}
-              searchQueryLanguage={searchQueryLanguage}
-              setSearchParams={setSearchParams}
-            />
-          </EuiFlexItem>
-          {docStats?.totalCount !== undefined && (
-            <EuiFlexItem>
-              <DocumentCountContent
-                brushSelectionUpdateHandler={setWindowParameters}
-                documentCountStats={docStats.documentCountStats}
-                totalCount={docStats.totalCount}
-              />
-            </EuiFlexItem>
-          )}
-          <EuiSpacer size="m" />
-          {earliest !== undefined && latest !== undefined && windowParameters !== undefined && (
-            <EuiFlexItem>
-              <ExplainLogRateSpikes
-                dataView={dataView}
-                earliest={earliest}
-                latest={latest}
-                windowParameters={windowParameters}
-              />
-            </EuiFlexItem>
-          )}
-        </EuiFlexGroup>
-      </EuiPageContentBody>
-    </EuiPageBody>
+    <UrlStateContextProvider value={{ searchString: urlSearchString, setUrlState }}>
+      <EuiPageBody data-test-subj="aiopsIndexPage" paddingSize="none" panelled={false}>
+        <ExplainLogRateSpikes dataView={dataView} />
+      </EuiPageBody>
+    </UrlStateContextProvider>
   );
 };
